requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.2
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary==2.9.9